Remove top-level CreateEvolvableToken object and Initiator/Responder flows in favour of top-level separated flows. Intended to ease developer use of the flows with simpler semantics.
